pyproject.toml
[project]
name = "resume-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.43.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
